Added infinite image scroll
